requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.8.4
aiohttp-retry==2.9.1
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.6.3
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.3.0
audioread==3.0.1
bitarray==2.9.3
black==23.3.0
cachetools==4.2.2
ccxt==4.4.62
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==1.0.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.1
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
decorator==5.2.1
eth-account==0.8.0
eth-hash==0.7.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
eth_abi==5.0.1
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gymnasium==1.0.0
h11==0.14.0
h5py==3.13.0
hexbytes==0.3.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
lru-dict==1.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==23.2
pandas==2.2.3
parsimonious==0.9.0
pathspec==0.12.1
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prometheus_client==0.21.1
propcache==0.3.0
protobuf==6.30.1
py-solc-x==2.0.3
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pyparsing==3.2.1
pytest==7.3.1
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2025.1
pyunormalize==16.0.0
pywin32==309
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.28.2
rich==13.9.4
rlp==3.0.0
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stable_baselines3==2.5.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.2.1
toolz==1.0.0
torch==2.6.0
tornado==6.1
twilio==9.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==1.26.20
web3==6.0.0
websocket-client==1.8.0
websockets==11.0.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
